+++ CHANGELOG.md
@@ -4,10 +4,15 @@ All notable changes to the "open-related-file" extension will be documented in t
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased] - 2018-09-15
+### Added
 - Add `openSingleFile` option that defaults to true.
 - If there is only one related file, open it without prompting (if `openSingleFile` is true).
 
+### Changed
+- Add "Open Related Files:" prefix to information messages shown by the extension.
+
 ## [0.0.3] - 2018-09-14
+### Added
 - Add better metadata
 
 ## [0.0.2] - 2018-09-14